[NORM]

+++ corewar_dir/srcs/ope/arithmetic_and_binary_ope_tool.c
@@ -6,7 +6,7 @@
 /*   By: ffoissey <ffoissey@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/17 12:07:18 by ffoissey          #+#    #+#             */
-/*   Updated: 2020/04/24 16:24:19 by ffoissey         ###   ########.fr       */
+/*   Updated: 2020/04/26 14:47:14 by ffoissey         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,15 +19,15 @@ static int8_t	process_op(t_carriages *current, enum e_type type, int32_t *arg)
 
 	flag_reg = SET | GET;
 	if (type == OP_AND)
-		reg_val = set_reg_value(current, (int8_t)arg[2], arg[0] & arg[1], &flag_reg);
+		reg_val = set_reg_value(current, arg[2], arg[0] & arg[1], &flag_reg);
 	else if (type == OP_OR)
-		reg_val = set_reg_value(current, (int8_t)arg[2], arg[0] | arg[1], &flag_reg);
+		reg_val = set_reg_value(current, arg[2], arg[0] | arg[1], &flag_reg);
 	else if (type == OP_XOR)
-		reg_val = set_reg_value(current, (int8_t)arg[2], arg[0] ^ arg[1], &flag_reg);
+		reg_val = set_reg_value(current, arg[2], arg[0] ^ arg[1], &flag_reg);
 	else if (type == OP_ADD)
-		reg_val = set_reg_value(current, (int8_t)arg[2], arg[0] + arg[1], &flag_reg);
+		reg_val = set_reg_value(current, arg[2], arg[0] + arg[1], &flag_reg);
 	else if (type == OP_SUB)
-		reg_val = set_reg_value(current, (int8_t)arg[2], arg[0] - arg[1], &flag_reg);
+		reg_val = set_reg_value(current, arg[2], arg[0] - arg[1], &flag_reg);
 	else
 		return (FAILURE);
 	if (flag_reg == BAD_REG)
